studymate/requirements.txt
streamlit>=1.36.0
pymupdf>=1.24.0
sentence-transformers>=2.7.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.1
requests>=2.32.0
tqdm>=4.66.0
pydantic>=2.7.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
rich>=13.7.0
uvicorn>=0.30.0
fastapi>=0.111.0
numpy>=1.26.0
scikit-learn>=1.5.0
openai>=1.30.0 # optional fallback
nltk>=3.9
spacy>=3.7.4
reportlab>=4.2.0
PyPDF2>=3.0.1
transformers>=4.41.0
torch>=2.2.0
